refactor(sandbox): separate removal command and service
- Keep CLI argument routing separate from removal coordination.
- Remove redundant cleanup candidate wrappers and unused exports.
- Point safety tests at the production service and filesystem boundary.
